Added ClockBroken()
basically when a error that renders the clock unusable happens this will get called and will place in the logs text on how to proceed with the clock being broken. Also put a yield break in UpdateTime to ensure it doesn't sit in an infinite loop when it isn't working.